requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.1.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==4.2
django-filter==24.2
django-celery-beat==2.6.0
django-cors-headers==4.7.0
django-cron==0.6.0
django-rest-passwordreset==1.5.0
django-timezone-field==6.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.6.0
holidays==0.36
idna==3.10
kombu==5.5.2
multidict==6.4.3
pillow==11.2.1
prettytable==3.16.0
prompt-toolkit==3.0.39
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.0.1
reportlab==4.4.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
twilio==9.6.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.6
yarl==1.20.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.2.0
